requirements-azure.txt
# PromoAgente - Azure Functions Requirements (Python 3.11)

# ========== Azure Functions Core ==========
azure-functions>=1.18.0
azure-functions-durable>=1.2.9

# ========== AI Framework ==========
agno==2.1.9
openai==1.40.0

# ========== Azure Services ==========
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# ========== Web Framework (para desenvolvimento local) ==========
fastapi==0.110.0
uvicorn==0.24.0

# ========== Utilities ==========
python-dotenv==1.0.0
aiosqlite==0.19.0
sqlalchemy==2.0.23
httpx<0.28

# ========== Templates and Forms ==========
jinja2==3.1.2
python-multipart==0.0.6

# ========== Data Processing ==========
openpyxl==3.1.2
pydantic>=2.0.0

# ========== Logging and Monitoring ==========
azure-monitor-opentelemetry>=1.2.0

# ========== JSON Processing ==========
orjson>=3.9.0
